[FIX] - Validation Period wurde als String 0.0 definiert und anschließend in ein Integer gecastet. Dies verursachte den Fehler "invalid literal for int() with base 10: '0.0'" Lösung: Typecast nun in ein Float.

+++ equitania/res_config.py
@@ -93,7 +93,7 @@ class sm_account_config_settings(osv.osv_memory):
          
         validation_period_str = config_parameter_obj.get_param(cr, uid, 'offer.valid.duration', context=context) or '0'
 
-        validation_period = int(validation_period_str)
+        validation_period = float(validation_period_str)
          
         return {
                 'eq_offer_valid_duration': validation_period,
